pimd: clarify some int variable conversions
Add explicit int types to some arithmetic expressions
where int overflow was possible.

+++ pimd/pim_cmd_common.c
@@ -4289,7 +4289,7 @@ void show_mroute_summary(struct pim_instance *pim, struct vty *vty,
 		json_object_int_add(json_starg, "installed",
 				    starg_hw_mroute_cnt);
 		json_object_int_add(json_starg, "total",
-				    starg_sw_mroute_cnt + starg_hw_mroute_cnt);
+				    (int64_t)starg_sw_mroute_cnt + starg_hw_mroute_cnt);
 
 		/* (S, G) route details */
 		json_sg = json_object_new_object();
@@ -4297,14 +4297,13 @@ void show_mroute_summary(struct pim_instance *pim, struct vty *vty,
 
 		json_object_int_add(json_sg, "installed", sg_hw_mroute_cnt);
 		json_object_int_add(json_sg, "total",
-				    sg_sw_mroute_cnt + sg_hw_mroute_cnt);
+				    (int64_t)sg_sw_mroute_cnt + sg_hw_mroute_cnt);
 
 		json_object_int_add(json, "totalNumOfInstalledMroutes",
-				    starg_hw_mroute_cnt + sg_hw_mroute_cnt);
+				    (int64_t)starg_hw_mroute_cnt + sg_hw_mroute_cnt);
 		json_object_int_add(json, "totalNumOfMroutes",
-				    starg_sw_mroute_cnt + starg_hw_mroute_cnt +
-					    sg_sw_mroute_cnt +
-					    sg_hw_mroute_cnt);
+				    (int64_t)starg_sw_mroute_cnt + starg_hw_mroute_cnt +
+					    sg_sw_mroute_cnt + sg_hw_mroute_cnt);
 	}
 }
